+++ CHANGELOG.md
@@ -19,6 +19,7 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 ### Fixes and improvements
 
 * Improve greedy decoding speed (up to 50% faster)
+* Fix error when building an inference graph including a `DenseBridge`
 
 ## [1.13.0](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.13.0) (2018-11-14)